data-pipeline/requirements.txt
pandas
yfinance
sqlalchemy
python-dotenv
psycopg2